refactored all type declarations into the primary reducer. All state should live together
